bookproject/requirements.txt
Django==4.2.24
python-dotenv==1.2.1
Requests==2.32.5
asgiref==3.8.1
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10